requirements.txt
discord.py==2.4.0
requests==2.32.3
bs4==0.0.2
lxml==4.9.3
python-dotenv==1.1.0